BED-6331: Added formatting function to display comma separated numbers (#1874)
* feat: added a new utils function and gave file a more generic name

* chore: updated import of util function after file name change
