+++ packages/javascript/bh-shared-ui/src/utils/index.ts
@@ -14,7 +14,6 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './abbreviatedNumber';
 export * from './api';
 export * from './array';
 export * from './colors';
@@ -28,6 +27,7 @@ export * from './exportGraphData';
 export * from './icons';
 export * from './jobs';
 export * from './luxon';
+export * from './numberFormatting';
 export * from './object';
 export * from './parseItemId';
 export * from './passwd';
